new feature: comment notifications

+++ pages/api/comments/[pageID].js
@@ -1,18 +1,11 @@
 import { getSession } from 'next-auth/client'
 import { ObjectId } from 'bson'
 import { connectToDatabase } from '../../../lib/db'
+import { ObjectID } from 'mongodb'
 
 export default async function handler(req, res) {
   const session = await getSession({ req })
   const pageID = req.query.pageID
-
-  if (!session) {
-    res.status(401).json({
-      message: 'No active session found.',
-      comments: [],
-    })
-    return
-  }
 
   if (!pageID) {
     res.status(400).json({
@@ -53,17 +46,56 @@ export default async function handler(req, res) {
     } else {
       res.status(201).json({ comments: comments })
     }
+
+    return
   }
 
   /* if POST method, create a new comment */
   if (req.method === 'POST') {
-    if (!session) res.status(401).json({ message: 'No active session found.' })
     const { username, text } = req.body
 
+    /* API protections - reject if no session, or if session name does not match
+    comment author's name */
+    if (!session || session.user.name !== username) {
+      res.status(401).json({ message: 'Invalid credentials.' })
+      return
+    }
     if (!username || !text || text.trim() === '' || text.trim().length > 500) {
       res.status(422).json({ message: 'Invalid input.' })
       return
     }
+
+    /* BEGIN NOTIFICATION SETTER */
+    /* first check if comment belongs to a user profile */
+    const usersCollection = db.collection('users')
+    const notifyUser = await usersCollection.findOne({
+      _id: ObjectId(pageID),
+    })
+    if (notifyUser) {
+      /* user profile comment has been commented upon, so let's increment
+      the notifs object in that user's document only, since only that user
+      needs the notification, using the user's own ID */
+      await usersCollection.updateOne(
+        { _id: ObjectId(pageID) },
+        { $inc: { notifs: 1 } }
+      )
+    } else {
+      /* no user found, so check if it's a show profile being commented upon */
+      const showsCollection = db.collection('shows')
+      const notifyShow = await showsCollection.findOne({
+        _id: ObjectId(pageID),
+      })
+      if (notifyShow) {
+        /* it is, so let's loop through each of the "excitedUsers", add each of them
+        as a property in the show document's notifs field, and then increment each one */
+        notifyShow.excitedUsers.forEach(async (user) => {
+          const notifField = { $inc: {} }
+          notifField['$inc']['notifs.' + user] = 1
+          await showsCollection.updateOne({ _id: ObjectId(pageID) }, notifField)
+        })
+      }
+    }
+    /* END NOTIFICATION SETTER */
 
     /* create a new comment object */
     const newComment = {
@@ -84,13 +116,20 @@ export default async function handler(req, res) {
     } else {
       res.status(201).json({ message: 'Comment created.' })
     }
+
+    return
   }
 
   /* if DELETE method, delete the selected comment */
   if (req.method === 'DELETE') {
-    if (!session) res.status(401).json({ message: 'No active session found.' })
     /* object destructuring */
-    const { commentID } = req.body
+    const { commentID, username } = req.body
+
+    /* only allow auth'd users to delete their own comments */
+    if (!session || session.user.name !== username) {
+      res.status(401).json({ message: 'Invalid credentials.' })
+      return
+    }
 
     /* Use the $pull operator to remove the ID'd comment from the comments array */
     const result = await commentsCollection.deleteOne({
